requirements.txt
torch==1.11.0
torchvision==0.12.0
numpy==1.23.0
pandas==2.0.3
pillow==9.4.0
matplotlib==3.6.2

